fix(ramp): Circle idempotencyKey UUID format and webhook string/number parsing
- Circle requires idempotencyKey to be a valid UUID; use order.TransactionID
  directly (already a UUID) for SOL path and uuid.NewSHA1 for EVM path
- RampHub webhook sends fiatAmount/cryptoAmount/exchangeRate as JSON strings,
  not numbers — add FlexFloat type that accepts both string and number
- Fix WebhookData field names/JSON tags to match actual RampHub webhook schema
  (platformUsed, cryptoSymbol, cryptoAmount, exchangeRate, blockchainTxHash)

+++ internal/domain/services/ramp/service.go
@@ -1021,9 +1021,9 @@ func (s *Service) HandleWebhook(ctx context.Context, deliveryID string, event *r
 	// Delivery-level idempotency is recorded only on success so a failed
 	// credit/reversal does not permanently consume the delivery ID — RampHub
 	// can retry and the work will be attempted again.
-	credErr := s.creditOnrampIfCompleted(ctx, userID, txID, newStatus, event.Data.TokenAmount, event.Data.FiatAmount, event.Data.TxHash)
+	credErr := s.creditOnrampIfCompleted(ctx, userID, txID, newStatus, float64(event.Data.TokenAmount), float64(event.Data.FiatAmount), event.Data.TxHash)
 	revErr := s.reverseOfframpIfFailed(ctx, userID, txID, orderType, newStatus)
-	s.notifyTerminal(ctx, userID, txID, orderType, newStatus, event.Data.FiatAmount)
+	s.notifyTerminal(ctx, userID, txID, orderType, newStatus, float64(event.Data.FiatAmount))
 	if credErr != nil {
 		return credErr
 	}
